[client] refactor wip
