+++ client/src/components/index.ts
@@ -3,3 +3,4 @@ export * from './Now';
 export * from './Player';
 export * from './SongModal';
 export * from './UserModal';
+export * from './SideMenu';
